feat: remove sync client, move & rename async client
